refactor: notifications group
